refactor: split monolithic data module into src/data/
